refactory: move conversion functions to own module in network crate
